develop: minor changes
